+++ src/libam_support/utils/mpi_utils.cc
@@ -15,7 +15,7 @@
 #include <string>
 
 namespace labw::art_modern {
-void exit_mpi(const int  /*status*/)
+void exit_mpi(const int /*status*/)
 {
     BOOST_LOG_TRIVIAL(info) << "EXIT";
 #ifdef WITH_MPI